database connection

+++ internal/types/types.go
@@ -1,8 +1,9 @@
 package types
 
 type Student struct {
-	Id    int    `json:"id"`
+	Id    int64   `json:"id"`
 	Name  string `json:"name" validate:"required"`
 	Email string `json:"email" validate:"required,email"`
 	Age   int    `json:"age" validate:"required"`
 }
+
